cql3: Remove term in update_statement
Replace all uses of term with expression in cql3/statements/update_statement

There was some trouble with extracting values from json.
The original code worked this way on a map example:
> There is a json string to parse: {'b': 1, 'a': 2, 'b': 3}
> The code parses the json and creates bytes where this map is serialized
  but without removing duplicates, sorting etc.
> Then a maps::delayed_value is created from these bytes.
  During creation map elements are extracted, sorted and duplicates are removed.
  This map value is then used in setter

Now when maps::delayed_value is changed to expr::constant the step where elements are sorted is lost.
Because of this we need to do this earlier, the best place is during original json parsing.

Additionally I suspect that removing duplicated elements used to work only on the first level, in case of map of maps it wouldn't work.
Now it will work no matter how many layers of maps there are.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>