cql3: expr: possible_lhs_values: Handle subscripted columns
Situation with possible_lhs_values and subscripted columns
is a bit wonky.

As its first argument possible_lhs_values takes a
column_definition for which it should find possible
values.

Sometimes we might want to find possible values
for a subscripted column, e.g `mycol[5]`.
This can be achieved by passing `mycol` as the first argument.

It's a bit of lie - we ask for values for `mycol`
and suddenly get values for `mycol[5]`, but I think
it's the most reasonable way things can be without
making deep changes in how possible_lhs_values looks like.
I think we can live with it.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>