prepare_expr: make it possible to prepare binary_operator using prepare_expression
prepare_expression didn't allow to prepare binary_operators.
Changed it so that it calls prepare_binary_operator.

This isn't the proper way to validate a query,
statement_restrictions combines preparing the binary operators
and validating them, but once they are validated we can
just use prepare_expression.

For now all validation is done inside statement_restrictions,
so after creating them we can assume that the binary operators
are valid and just use prepare_expression to get a prepared
WHERE clause.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>