expression: handle IS NOT NULL in is_satisfied_by
is_satisied_by wasn't able to handle IS NOT NULL
restrictions, so a handler was added.

Currently IS NOT NULL restrictions are handled
in a special way, but soon we will switch to
treating them like any other restriction.

is_satisfied_by should be able to handle them,
otherwise filtering of expressions with IS NOT NULL
will fail with on_internal_error.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>