expr: make it possible to evaluate binary_operator
evaluate() takes an expression and evaluates it
to a constant value. It wasn't possible to evalute
binary operators before, so it's added.
This is needed in prepare_expression: when preparing
a binary_operator that has constant on both
sides of the equation we would like to immediately
evalute it to a single value.

It looks like is_satisfied_by and evalate()
try do do the same thing, one could be implemented
using the other. In the future they might get
merged into a single function.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>